add `-latest` flag
